rename stubs file in demos
